feat(build): provide a lite build
Now providing a lite browser build:
- search only
- vanilla JS only (no angular, jquery lite builds for now)
- same export name (`algoliasearch`)
- now 5.93kb gzipped (versus 20kb)
- IE11/10/9 not natively supported, use http://cdn.polyfill.io/v2/docs/
  or just provide the Promise polyfill in the global object (window)
- same init usage
- has
  client.search/index.search/clearCache/setRequestTimeout/setExtraHeader/..
- no write/indexing/waitTask/settings methods => search only
- available on jsDelivr using algoliasearchLite[.min].js
- also requireable using
  var algoliasearch = require('algoliasearch/dist/algoliasearchLite.js');

fixes #215
