+++ .gitattributes
@@ -5,3 +5,5 @@ dist/algoliasearch.angular.min.js -diff
 dist/algoliasearch.jquery.js -diff
 dist/algoliasearch.jquery.min.js -diff
 dist/algoliasearch.parse.js -diff
+dist/algoliasearchLite.js -diff
+dist/algoliasearchLite.min.js -diff
