+++ examples/lite.html
@@ -1,0 +1,22 @@
+<!doctype html>
+<html>
+<head>
+  <title>AlgoliaSearch JavaScript Lite client</title>
+  <script src="/dist/algoliasearchLite.js"></script>
+</head>
+<body>
+  <script>
+    var client = algoliasearch('latency', '6be0576ff61c053d5f9a3225e2a90f76');
+    var index = client.initIndex('contacts');
+
+    index.search('Atlenta', function(err, results) {
+      if (err) {
+        throw err;
+      }
+
+      console.log('We got `' + results.nbHits + '` results');
+      console.log('Here is the first one: ', results.hits[0]);
+    });
+  </script>
+</body>
+</html>
