+++ src/browser/builds/lite/clone.js
@@ -1,0 +1,3 @@
+module.exports = function clone(obj) {
+  return JSON.parse(JSON.stringify(obj));
+};
