+++ src/browser/builds/lite/debug.js
@@ -1,0 +1,3 @@
+module.exports = function debug() {
+  return function noop() {};
+};
