+++ src/browser/builds/lite/forEach.js
@@ -1,0 +1,3 @@
+module.exports = function forEach(arr, fn) {
+  return arr.forEach(fn);
+};
