+++ src/browser/builds/lite/isArray.js
@@ -1,0 +1,3 @@
+module.exports = function isArray(arr) {
+  return Array.isArray(arr);
+};
