+++ src/browser/builds/lite/map.js
@@ -1,0 +1,3 @@
+module.exports = function map(arr, fn) {
+  return arr.map(fn);
+};
